Convert DocumentVersionCache to interface
